+++ src/plone/restapi/serializer/dxcontent.py
@@ -33,48 +33,59 @@ class SerializeToJson(object):
 
         self.permission_cache = {}
 
-    def __call__(self):
-        parent = aq_parent(aq_inner(self.context))
+    def getVersion(self, version_id):
+        if version_id == 'current':
+            return self.context
+        else:
+            repo_tool = getToolByName(self.context, "portal_repository")
+            return repo_tool.retrieve(self.context, int(version_id)).object
+
+    def __call__(self, version_id=None):
+        version_id = 'current' if version_id is None else version_id
+
+        obj = self.getVersion(version_id)
+        parent = aq_parent(aq_inner(obj))
         parent_summary = getMultiAdapter(
             (parent, self.request), ISerializeToJsonSummary)()
         result = {
             # '@context': 'http://www.w3.org/ns/hydra/context.jsonld',
-            '@id': self.context.absolute_url(),
-            'id': self.context.id,
-            '@type': self.context.portal_type,
+            '@id': obj.absolute_url(),
+            'id': obj.id,
+            '@type': obj.portal_type,
             'parent': parent_summary,
-            'created': json_compatible(self.context.created()),
-            'modified': json_compatible(self.context.modified()),
-            'review_state': self._get_workflow_state(),
-            'UID': self.context.UID(),
+            'created': json_compatible(obj.created()),
+            'modified': json_compatible(obj.modified()),
+            'review_state': self._get_workflow_state(obj),
+            'UID': obj.UID(),
+            'version_id': version_id,
             'layout': self.context.getLayout(),
         }
 
-        for schema in iterSchemata(self.context):
+        for schema in iterSchemata(obj):
 
             read_permissions = mergedTaggedValueDict(
                 schema, READ_PERMISSIONS_KEY)
 
             for name, field in getFields(schema).items():
 
-                if not self.check_permission(read_permissions.get(name)):
+                if not self.check_permission(read_permissions.get(name), obj):
                     continue
 
                 serializer = queryMultiAdapter(
-                    (field, self.context, self.request),
+                    (field, obj, self.request),
                     IFieldSerializer)
                 value = serializer()
                 result[json_compatible(name)] = value
 
         return result
 
-    def _get_workflow_state(self):
+    def _get_workflow_state(self, obj):
         wftool = getToolByName(self.context, 'portal_workflow')
         review_state = wftool.getInfoFor(
-            ob=self.context, name='review_state', default=None)
+            ob=obj, name='review_state', default=None)
         return review_state
 
-    def check_permission(self, permission_name):
+    def check_permission(self, permission_name, obj):
         if permission_name is None:
             return True
 
@@ -86,7 +97,7 @@ class SerializeToJson(object):
             else:
                 sm = getSecurityManager()
                 self.permission_cache[permission_name] = bool(
-                    sm.checkPermission(permission.title, self.context))
+                    sm.checkPermission(permission.title, obj))
         return self.permission_cache[permission_name]
 
 
@@ -100,8 +111,10 @@ class SerializeFolderToJson(SerializeToJson):
                  'sort_on': 'getObjPositionInParent'}
         return query
 
-    def __call__(self):
-        folder_metadata = super(SerializeFolderToJson, self).__call__()
+    def __call__(self, version_id=None):
+        folder_metadata = super(SerializeFolderToJson, self).__call__(
+            version_id=version_id
+        )
 
         query = self._build_query()