+++ src/plone/restapi/tests/test_services_content.py
@@ -1,0 +1,63 @@
+# -*- coding: utf-8 -*-
+from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
+from plone.restapi.testing import RelativeSession
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.app.textfield.value import RichTextValue
+from plone.namedfile.file import NamedBlobFile
+from plone.namedfile.file import NamedBlobImage
+
+from z3c.relationfield import RelationValue
+from zope.component import getUtility
+from zope.intid.interfaces import IIntIds
+
+import unittest
+import transaction
+
+
+class TestHistoryVersioning(unittest.TestCase):
+
+    layer = PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer['app']
+        self.portal = self.layer['portal']
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+
+        self.api_session = RelativeSession(self.portal_url)
+        self.api_session.headers.update({'Accept': 'application/json'})
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+        self.portal.invokeFactory(
+            'Document',
+            id='doc_with_history',
+            title='My Document'
+        )
+        self.doc = self.portal.doc_with_history
+        self.doc.setTitle('Current version')
+
+        transaction.commit()
+
+    def test_default_current(self):
+        # If we specify no version_id, do we get the current version?
+        response = self.api_session.get(self.doc.absolute_url())
+        self.assertEqual(response.json()['title'], 'Current version')
+
+    def test_explicit_current(self):
+        # Does version_id=current get the current version
+        url = self.doc.absolute_url() + '?version_id=current'
+        response = self.api_session.get(url)
+        self.assertEqual(response.json()['title'], 'Current version')
+
+    def test_previous_version(self):
+        # Does version_id=0 get the older version?
+        url = self.doc.absolute_url() + '?version_id=0'
+        response = self.api_session.get(url)
+        self.assertEqual(response.json()['title'], 'My Document')
+
+    def test_response(self):
+        response = self.api_session.get(self.doc.absolute_url())
+        self.assertIn('version_id', response.json())