+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.0a14 (unreleased)
+1.0a14 (2017-05-02)
 -------------------
 
 New Features: